Fixed a few more things

+++ DiscreteFrechet.cpp
@@ -36,7 +36,7 @@ double pointL2(double x1, double y1, double x2, double y2)
 }
 
 //void discreteFrechet(double **table, int i, int j, vector<tuple<int,double>>& v1, vector<tuple<int,double>>& v2) //how_many_columns -1 
-void discreteFrechet(double **table, int i, int j, vector<tuple<unsigned int,double>>& v1, vector<tuple<unsigned int,double>>& v2) //how_many_columns -1 
+void discreteFrechet(double **table, int i, int j, vector<tuple<double,double>>& v1, vector<tuple<double,double>>& v2) //how_many_columns -1 
 {
     //double table[i][j]{-1.0};
 
@@ -49,38 +49,32 @@ void discreteFrechet(double **table, int i, int j, vector<tuple<unsigned int,dou
     
     for (int l = 0; l < j; l++)
     {
-       
         double x1 = get<0>(v1[0]);
         double y1 = get<1>(v1[0]);
         double x2 = get<0>(v2[l]);
         double y2 = get<1>(v2[l]);
         double temp = pointL2(x1,y1,x2,y2); //abs(v1[0] - v2[l]);
-        
         if (l == 0)
         { //first row first column
             table[0][0] = c;
         }else{
             table[0][l] = max(table[0][l - 1], temp); //first row (except the first column)
-        }
-        
+        }    
     }
 
-    // for (int k = 0; k < j; k++)
+    // for (int k = 0; k < i; k++)
     // {
-       
     //     double x1 = get<0>(v1[k]);
     //     double y1 = get<1>(v1[k]);
     //     double x2 = get<0>(v2[0]);
     //     double y2 = get<1>(v2[0]);
     //     double temp = pointL2(x1,y1,x2,y2); //abs(v1[0] - v2[l]);
-        
     //     if (k == 0)
     //     { //first row first column
     //         table[0][0] = c;
     //     }else{
     //         table[k][0] = max(table[k - 1][0], temp); //first row (except the first column)
     //     }
-        
     // }
 
     for (int row = 1; row < i; row++) //for the rest of the table(second row every column)
@@ -124,7 +118,7 @@ void discreteFrechet(double **table, int i, int j, vector<tuple<unsigned int,dou
 }
 
 //double ret_DFD(int i, int j, vector<tuple<int,double>>& v1, vector<tuple<int,double>>& v2){
-double ret_DFD(int i, int j, vector<tuple<unsigned int,double>>& v1, vector<tuple<unsigned int,double>>& v2){
+double ret_DFD(int i, int j, vector<tuple<double,double>>& v1, vector<tuple<double,double>>& v2){
 
 
     double **table = new double*[i];
@@ -165,7 +159,7 @@ int minIndex(double x, double y, double z)
 }
 
 //void optimalTraversal(vector<tuple <int, double,int, double > >& traversal, double **table, int m1, int m2, std::vector<std::tuple<int,double>>& v1, std::vector<std::tuple<int,double>>& v2){
-void optimalTraversal(vector<tuple <unsigned int, double,unsigned int, double > >& traversal, double **table, int m1, int m2, std::vector<std::tuple<unsigned int,double>>& v1, std::vector<std::tuple<unsigned int,double>>& v2){
+void optimalTraversal(vector<tuple <double, double,double, double > >& traversal, double **table, int m1, int m2, std::vector<std::tuple<double,double>>& v1, std::vector<std::tuple<double,double>>& v2){
 
 
     int index;
@@ -246,7 +240,7 @@ CurveElement* MeanCurve(CurveElement* P, CurveElement* Q){
     discreteFrechet(table,i,i,P->arrayElementTwoD,Q->arrayElementTwoD);         
 
     //vector<tuple < int, double, int, double > > traversal;
-    vector<tuple < unsigned int, double, unsigned int, double > > traversal;
+    vector<tuple < double, double, double, double > > traversal;
     optimalTraversal(traversal,table,i,i,P->arrayElementTwoD,Q->arrayElementTwoD);
     //myLogFile << traversal.size() << endl;
