+++ DiscreteFrechet.h
@@ -12,12 +12,12 @@ extern std::ofstream myLogFile;
 
 double min_3(double x, double y, double z);
 //void discreteFrechet(double **table, int i, int j, std::vector<std::tuple<int,double>>& v1, std::vector<std::tuple<int,double>>& v2);
-void discreteFrechet(double **table, int i, int j, std::vector<std::tuple<unsigned int,double>>& v1, std::vector<std::tuple<unsigned int,double>>& v2);
+void discreteFrechet(double **table, int i, int j, std::vector<std::tuple<double,double>>& v1, std::vector<std::tuple<double,double>>& v2);
 //double ret_DFD(int i, int j, std::vector<std::tuple<int,double>>& v1, std::vector<std::tuple<int,double>>& v2);
-double ret_DFD(int i, int j, std::vector<std::tuple<unsigned int,double>>& v1, std::vector<std::tuple<unsigned int,double>>& v2);
+double ret_DFD(int i, int j, std::vector<std::tuple<double,double>>& v1, std::vector<std::tuple<double,double>>& v2);
 int minIndex(double x, double y, double z);
 //void optimalTraversal(std::vector<std::tuple <int, double,int, double > >& traversal,double **table, int m1, int m2, std::vector<std::tuple<int,double>>& v1, std::vector<std::tuple<int,double>>& v2);
-void optimalTraversal(std::vector<std::tuple <unsigned int, double,unsigned int, double > >& traversal,double **table, int m1, int m2, std::vector<std::tuple<unsigned int,double>>& v1, std::vector<std::tuple<unsigned int,double>>& v2);
+void optimalTraversal(std::vector<std::tuple <double, double, double, double > >& traversal,double **table, int m1, int m2, std::vector<std::tuple<double,double>>& v1, std::vector<std::tuple<double,double>>& v2);
 CurveElement* MeanCurve(CurveElement*, CurveElement*);
 
 #endif /* DISCRETEFRECHET_H */
